+++ CHANGELOG.md
@@ -31,6 +31,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - channel type input for slash options
 
+### Fixed
+
+- command update, when channel types are same
+
 ## [8.1.0] - 2021-11-03
 
 ### Added