+++ CHANGELOG.md
@@ -21,11 +21,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 `[FULL CHANGELOG](https://github.com/oceanroleplay/discord.ts/blob/main/CHANGELOG.md#820---2021-11-10)`
 
-## [8.2.0] - 2021-11-10
+## [9.0.0] - 2021-11-10
 
 ### Added
 
 - regex support for button/selectmenu decorator
+- ESM Support
 
 ### Changed